Made npm run seed available; made generated prices more reasonable; small css changes
